Change Link to Navlink in Navbar. Add active-link class to underline the active tab
