+++ src/App.js
@@ -1,8 +1,8 @@
 import './App.scss';
-// import { Provider } from 'react-redux';
+import { Provider } from 'react-redux';
+import store from './store';
 // import { createLogger } from 'redux-logger';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-// import { Provider } from 'react-redux';
 import Rockets from './components/Rockets/Rockets';
 import Profile from './components/Profile/Profile';
 import Missions from './components/Missions/Missions';
@@ -10,7 +10,7 @@ import Navbar from './components/Navbar/Navbar';
 
 function App() {
   return (
-  // <Provider>
+  <Provider store={store}>
     <div className="wrapper">
       <BrowserRouter>
         <Navbar />
@@ -23,7 +23,7 @@ function App() {
         </div>
       </BrowserRouter>
     </div>
-  // </Provider>
+  </Provider>
   );
 }
