--- src/components/Missions/Missions.jsx
@@ -1,11 +1,0 @@
-const Missions = () => (
-  <div className="missions">
-    <h1>Missions</h1>
-    <p>
-      Lorem ipsum dolor sit amet consectetur adipisicing elit.
-      officiis dicta ipsaui.
-    </p>
-  </div>
-);
-
-export default Missions;
