--- src/components/Navbar/Navbar.jsx
@@ -1,24 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import './Navbar.scss';
-
-const Navbar = () => (
-  <nav>
-    <span className="logo">
-      Space Travelers&rsquo; Hub
-    </span>
-    <ul>
-      <li>
-        <Link to="/">Rockets</Link>
-      </li>
-      <li>
-        <Link to="/missions">Missions</Link>
-      </li>
-      <li>
-        <Link to="/profile">Profile</Link>
-      </li>
-    </ul>
-  </nav>
-);
-
-export default Navbar;
