--- src/components/Profile/Profile.jsx
@@ -1,11 +1,0 @@
-const Profile = () => (
-  <>
-    <h1>Profile</h1>
-    <p>
-      Lorem ipsum dolor sit amet consectetur adipisicing elit.
-      officiis dicta ipsaui.
-    </p>
-  </>
-);
-
-export default Profile;
