--- src/components/Rockets/Rockets.jsx
@@ -1,11 +1,0 @@
-const Rockets = () => (
-  <>
-    <h1>Rockets</h1>
-    <p>
-      Lorem ipsum dolor sit amet consectetur adipisicing elit.
-      officiis dicta ipsaui.
-    </p>
-  </>
-);
-
-export default Rockets;
