+++ src/views/Missions/Missions.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const Missions = () =>
+  (
+    <>
+      <h1>Missions</h1>
+      <p>
+        Lorem ipsum dolor sit amet consectetur adipisicing elit.
+        officiis dicta ipsaui.
+      </p>
+    </>
+  );
+
+export default Missions;
