+++ src/views/Navbar/Navbar.scss
@@ -1,0 +1,12 @@
+.navbar {
+
+  ul {
+
+    li {
+
+      .active-link {
+        text-decoration: underline;
+      }
+    }
+  }
+}
