+++ src/views/Profile/Profile.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Profile = () => (
+    <>
+      <h1>Profile</h1>
+      <p>
+        Lorem ipsum dolor sit amet consectetur adipisicing elit.
+        officiis dicta ipsaui.
+      </p>
+    </>
+);
+
+export default Profile;
