+++ src/views/Rockets/Rockets.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Rockets = () => (
+    <div className="add-book">
+      <h1>Rockets</h1>
+      <p>
+        Lorem ipsum dolor sit amet consectetur adipisicing elit.
+        officiis dicta ipsaui.
+      </p>
+    </div>
+);
+
+export default Rockets;
